test(gateway): focus timeout fallback coverage
